+++ spec/integration/braintree/transaction_gateway_spec.js
@@ -2306,7 +2306,7 @@ describe('TransactionGateway', function () {
     );
 
     context('with a local payment', function () {
-      it('returns LocalPayment for payment_instrument', done =>
+      it('returns relevant local payment transaction details for a sale', done =>
         specHelper.defaultGateway.customer.create({}, function () {
           let transactionParams = {
             paymentMethodNonce: Nonces.LocalPayment,
@@ -2323,6 +2323,43 @@ describe('TransactionGateway', function () {
             assert.isString(response.transaction.localPayment.payerId);
             assert.isString(response.transaction.localPayment.paymentId);
             assert.isString(response.transaction.localPayment.fundingSource);
+            assert.isString(response.transaction.localPayment.captureId);
+            assert.isString(response.transaction.localPayment.debugId);
+            assert.isNotNull(response.transaction.localPayment.transactionFeeAmount);
+            assert.isString(response.transaction.localPayment.transactionFeeCurrencyIsoCode);
+
+            done();
+          });
+        })
+      );
+
+      it('returns relevant local payment transaction details for a refund', done =>
+        specHelper.defaultGateway.customer.create({}, function () {
+          let transactionParams = {
+            paymentMethodNonce: Nonces.LocalPayment,
+            amount: '100.00',
+            options: {
+              submitForSettlement: true
+            }
+          };
+
+          specHelper.defaultGateway.transaction.sale(transactionParams, function (err, response) {
+            assert.isNull(err);
+            assert.isTrue(response.success);
+
+            specHelper.defaultGateway.transaction.refund(response.transaction.id, function (err, response) {
+              assert.isNull(err);
+              assert.isTrue(response.success);
+
+              assert.equal(response.transaction.paymentInstrumentType, PaymentInstrumentTypes.LocalPayment);
+              assert.isString(response.transaction.localPayment.payerId);
+              assert.isString(response.transaction.localPayment.paymentId);
+              assert.isString(response.transaction.localPayment.fundingSource);
+              assert.isString(response.transaction.localPayment.refundId);
+              assert.isString(response.transaction.localPayment.debugId);
+              assert.isNotNull(response.transaction.localPayment.refundFromTransactionFeeAmount);
+              assert.isString(response.transaction.localPayment.refundFromTransactionFeeCurrencyIsoCode);
+            });
 
             done();
           });